task: redirect user to last filled step
When a user has filled a certain part of the form and then leave the page
to be revisited later, the user is redirected to the last part of the
form that was filled provided the user clicked the next button before
leaving the page. The user data is then pre-filled with the data they
last entered.

Data are stored in the local storage of the browser, and if the user at
any point in time cleared the it saved data will be lost, and user will
have to start all over again.

Also worth noting is that the local storage is a temporary measure,it's
a security risk and should not be relied on on a production environment.